requirement.txt
pillow
numpy
wand
pypiwin32
